src/web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.9",
    "djangorestframework>=3.16.1",
    "django-cors-headers==4.9.0",
    "django-health-check>=4.0.6",
    "psycopg2-binary==2.9.11",
    "algoliasearch-django>=4.0.0",
    "algoliasearch>=4.35.4",
    "python-dotenv>=1.2.1",
    "requests==2.32.5",
    "bcrypt==5.0.0",
    "passlib",
    "python-jose",
    "drf-spectacular>=0.29.0",
    "django-debug-toolbar>=6.2.0",
    "dotenv>=0.9.9",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_test.py", "tests_*.py"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "config*"]
exclude = ["tests*", "client*", "templates*", "reports*", "build*", "dist*", "*.egg-info*", "__pycache__*"]
